Update example-node-operation to OF 0.9.0
