+++ example-node-operation/addons.make
@@ -1,0 +1,3 @@
+ofxCv
+ofxInteractivePrimitives
+ofxOpenCv
